Devided bot class
