--- src/telegram_bot/bot_registration.py
@@ -1,6 +1,0 @@
-from src.telegram_bot.base_bot import BaseBot
-
-
-class BotRegistration(BaseBot):
-    pass
-
