+++ src/telegram_bot/resources/actions.py
@@ -2,7 +2,7 @@ class Actions(object):
     # Actions
     FREE_TRIAL = 'free_trial'
     TOOLS = 'tools'
-    BACK  = 'back'
+    BACK = 'back'
     INFO = 'info'
     ALERTS = 'alerts'
     DILUTION = 'dilution'
